+++ DCManagement/functions/AccessRules/Test-DCAccessRule.ps1
@@ -32,6 +32,9 @@
 		
 		[PSCredential]
 		$Credential,
+
+        [string[]]
+        $TargetServer,
 		
 		[switch]
 		$EnableException
@@ -40,6 +43,9 @@
 	begin
 	{
 		$parameters = $PSBoundParameters | ConvertTo-PSFHashtable -Include Server, Credential
+        if (-not $Server -and $TargetServer) {
+            $parameters.Server = $TargetServer | Select-Object -First 1
+        }
 		$parameters['Debug'] = $false
 		Assert-ADConnection @parameters -Cmdlet $PSCmdlet
 		Invoke-PSFCallback -Data $parameters -EnableException $true -PSCmdlet $PSCmdlet
@@ -225,6 +231,7 @@
 	{
 		foreach ($domainController in $domainControllers)
 		{
+            if ($TargetServer -and $domainController.Name -notin $TargetServer) { continue }
 			$results = @{
 				ObjectType = 'FSAccessRule'
 				Server	   = $domainController.Name