+++ DCManagement/functions/shares/Invoke-DCShare.ps1
@@ -40,6 +40,9 @@
 		
 		[PSFComputer]
 		$Server,
+
+        [string[]]
+        $TargetServer,
 		
 		[PSCredential]
 		$Credential,
@@ -51,6 +54,9 @@
 	begin
 	{
 		$parameters = $PSBoundParameters | ConvertTo-PSFHashtable -Include Server, Credential
+        if (-not $Server -and $TargetServer) {
+            $parameters.Server = $TargetServer | Select-Object -First 1
+        }
 		$parameters['Debug'] = $false
 		Assert-ADConnection @parameters -Cmdlet $PSCmdlet
 		Invoke-PSFCallback -Data $parameters -EnableException $true -PSCmdlet $PSCmdlet
@@ -63,6 +69,7 @@
 	}
 	process
 	{
+        if ($TargetServer) { $parameters.TargetServer = $TargetServer }
 		if (-not $InputObject) { $InputObject = Test-DCShare @parameters }
 		
 		foreach ($testItem in $InputObject)